+++ src/main/java/com/turing/forseason/domain/user/controller/login/AuthController.java
@@ -8,17 +8,16 @@ import com.turing.forseason.domain.user.dto.SignInRequestDto;
 import com.turing.forseason.domain.user.dto.SignUpRequestDto;
 import com.turing.forseason.domain.user.dto.UserDetailDto;
 import com.turing.forseason.domain.user.entity.UserEntity;
+import com.turing.forseason.domain.user.service.GeneralAuthService;
+import com.turing.forseason.domain.user.service.KakaoAuthService;
 import com.turing.forseason.domain.user.service.UserService;
 import com.turing.forseason.global.dto.ApplicationResponse;
 import com.turing.forseason.global.errorException.ErrorCode;
-import com.turing.forseason.global.jwt.JwtTokenProvider;
+import com.turing.forseason.global.jwt.JwtTokenDto;
 import com.turing.forseason.global.jwt.OauthToken;
 import com.turing.forseason.global.jwt.PrincipalDetails;
-import com.turing.forseason.global.redis.RedisService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.parameters.P;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
@@ -26,22 +25,23 @@ import javax.servlet.http.HttpServletRequest;
 @CrossOrigin(origins = "*")
 @RestController
 @RequiredArgsConstructor
-public class UserController {
+public class AuthController {
     private final UserService userService;
-    private final RedisService redisService;
+    private final GeneralAuthService generalAuthService;
+    private final KakaoAuthService kakaoAuthService;
 
     @GetMapping("/api/login/oauth2/code/kakao")
-    public ApplicationResponse<String> Login(@RequestParam("code") String code) {
+    public ApplicationResponse<JwtTokenDto> Login(@RequestParam("code") String code) {
 
         // 카카오로부터 OauthToken 발급받기
-        OauthToken oauthToken = userService.getKakaoAccessToken(code);
+        OauthToken oauthToken = kakaoAuthService.getKakaoAccessToken(code);
         // 발급 받은 OauthToken 으로 카카오 회원 정보 DB 저장 후 JWT 를 생성
         System.out.println(oauthToken);
-        String jwtToken = userService.saveUserAndGetToken(oauthToken);
+        JwtTokenDto jwtTokenDto = kakaoAuthService.saveUserAndGetToken(oauthToken);
         System.out.println("jwt 토큰 발급");
-        System.out.println(jwtToken);
+        System.out.println(jwtTokenDto);
 
-        return ApplicationResponse.ok(ErrorCode.SUCCESS_CREATED, JwtTokenProvider.TOKEN_PREFIX + jwtToken);
+        return ApplicationResponse.ok(ErrorCode.SUCCESS_CREATED, jwtTokenDto);
     }
 
     @GetMapping("/auth/me")
@@ -76,7 +76,7 @@ public class UserController {
     public ApplicationResponse<String> serviceLogout(@AuthenticationPrincipal PrincipalDetails principalDetails) {
         // 서비스 로그아웃 (토큰 만료시키기)
         System.out.println("서비스 로그아웃");
-        userService.serviceLogout(principalDetails);
+        kakaoAuthService.serviceLogout(principalDetails);
         return ApplicationResponse.ok(ErrorCode.SUCCESS_OK, "서비스 로그아웃 되었습니다.");
     }
 
@@ -84,14 +84,14 @@ public class UserController {
     // 여기서부터는 일반로그인
     @GetMapping("/signup/verification/email")
     public ApplicationResponse<String> getEmailAuthCode(@RequestParam(name = "email") String email) {
-        if(userService.isDuplicatedEmail(email))
-            userService.sendEmailAuthCode(email);
+        if(generalAuthService.isDuplicatedEmail(email))
+            generalAuthService.sendEmailAuthCode(email);
         return ApplicationResponse.ok(ErrorCode.SUCCESS_OK, "인증 코드가 전송되었습니다.");
     }
 
     @PostMapping("/signup/verification/email")
     public ApplicationResponse<String> verifyEmail(@RequestBody EmailVerificationDto emailVerificationDto) {
-        userService.verifyEmail(emailVerificationDto);
+        generalAuthService.verifyEmail(emailVerificationDto);
         return ApplicationResponse.ok(ErrorCode.SUCCESS_OK, "인증에 성공하였습니다.");
     }
 
@@ -99,17 +99,17 @@ public class UserController {
     public ApplicationResponse<String> generalSignUp(@RequestBody SignUpRequestDto requestDto) {
         // 코드 리팩토링 필수(URI 설정, 비밀번호 암호화 등)
         System.out.println("일반 회원가입");
-        userService.signUpUser(requestDto);
+        generalAuthService.signUpUser(requestDto);
         return ApplicationResponse.ok(ErrorCode.SUCCESS_CREATED, "회원가입 성공");
     }
 
     @PostMapping("/signin/general")
-    public ApplicationResponse<String> generalSignIn(@RequestBody SignInRequestDto requestDto) {
+    public ApplicationResponse<JwtTokenDto> generalSignIn(@RequestBody SignInRequestDto requestDto) {
         System.out.println("일반 로그인");
-        String jwtToken = userService.signInAndGetToken(requestDto);
-        System.out.println(jwtToken);
+        JwtTokenDto jwtTokenDto = generalAuthService.signInAndGetToken(requestDto);
+        System.out.println(jwtTokenDto);
 
-        return ApplicationResponse.ok(ErrorCode.SUCCESS_OK, JwtTokenProvider.TOKEN_PREFIX + jwtToken);
+        return ApplicationResponse.ok(ErrorCode.SUCCESS_OK, jwtTokenDto);
     }
 
 
@@ -117,7 +117,7 @@ public class UserController {
     @GetMapping("test/oauthtoken/expried")
     public ApplicationResponse<String> test1(@AuthenticationPrincipal PrincipalDetails principalDetails) {
         UserEntity user = principalDetails.getUser();
-        boolean isExpired = userService.isExpired(user.getUserId());
+        boolean isExpired = kakaoAuthService.isExpired(user.getUserId());
 
         return ApplicationResponse.ok(ErrorCode.SUCCESS_OK, "" + isExpired);
     }
@@ -125,7 +125,7 @@ public class UserController {
     @GetMapping("test/oauthtoken/refresh")
     public ApplicationResponse<OauthToken> test2(@AuthenticationPrincipal PrincipalDetails principalDetails) {
         UserEntity user = principalDetails.getUser();
-        OauthToken newOauthToken = userService.getRefresh(user.getUserId());
+        OauthToken newOauthToken = kakaoAuthService.getRefresh(user.getUserId());
 
         return ApplicationResponse.ok(ErrorCode.SUCCESS_OK, newOauthToken);
     }
@@ -133,9 +133,9 @@ public class UserController {
     @GetMapping("test/oauthtoken/profile")
     public ApplicationResponse<KakaoProfile> test3(@AuthenticationPrincipal PrincipalDetails principalDetails) {
         UserEntity user = principalDetails.getUser();
-        OauthToken oauthToken = userService.getOauthToken(user.getUserId());
+        OauthToken oauthToken = kakaoAuthService.getOauthToken(user.getUserId());
 
-        KakaoProfile kakaoProfile = userService.findProfile(oauthToken);
+        KakaoProfile kakaoProfile = kakaoAuthService.findProfile(oauthToken);
 
         return ApplicationResponse.ok(ErrorCode.SUCCESS_OK, kakaoProfile);
     }
