+++ src/main/java/com/turing/forseason/global/jwt/JwtTokenDto.java
@@ -1,0 +1,19 @@
+package com.turing.forseason.global.jwt;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class JwtTokenDto {
+    private String accessToken;
+    private String refreshToken;
+
+    @Builder
+    public JwtTokenDto(String accessToken, String refreshToken) {
+        this.accessToken = accessToken;
+        this.refreshToken = refreshToken;
+    }
+}
